+++ xclim/indices/_agro.py
@@ -5,9 +5,10 @@ from typing import Optional
 import xarray
 
 import xclim.indices as xci
-from xclim.core.units import convert_units_to, declare_units
+import xclim.indices.run_length as rl
+from xclim.core.units import convert_units_to, declare_units, rate2amount, to_agg_units
 from xclim.core.utils import DayOfYearStr
-from xclim.indices.generic import aggregate_between_dates
+from xclim.indices.generic import aggregate_between_dates, compare
 
 # Frequencies : YS: year start, QS-DEC: seasons starting in december, MS: month start
 # See http://pandas.pydata.org/pandas-docs/stable/timeseries.html#offset-aliases
@@ -17,9 +18,12 @@ from xclim.indices.generic import aggregate_between_dates
 # -------------------------------------------------- #
 
 __all__ = [
-    "corn_heat_units",
     "biologically_effective_degree_days",
     "cool_night_index",
+    "corn_heat_units",
+    "dry_spell_frequency",
+    "dry_spell_total_length",
+    "latitude_temperature_index",
     "water_budget",
 ]
 
@@ -188,7 +192,7 @@ def biologically_effective_degree_days(
     .. math::
         k = f(lat) = 1 + \left(\frac{\left| lat  \right|}{50} * 0.06,  \text{if }40 < |lat| <50, \text{else } 0\right)
 
-    A second version of the BEDD (`method="icclim") does not consider :math:`TR_{adj}` and :math:`k` and employs a
+    A second version of the BEDD (`method="icclim"`) does not consider :math:`TR_{adj}` and :math:`k` and employs a
     different end date (30 September). The simplified formula is as follows:
 
     .. math::
@@ -287,6 +291,68 @@ def cool_night_index(
     return cni
 
 
+@declare_units(tas="[temperature]")
+def latitude_temperature_index(
+    tas: xarray.DataArray,
+    lat: xarray.DataArray,
+    lat_factor: float = 75,
+    freq: str = "YS",
+) -> xarray.DataArray:
+    """Latitude-Temperature Index.
+
+    Mean temperature of the warmest month with a latitude-based scaling factor.
+    Used for categorizing winegrowing regions.
+
+    Parameters
+    ----------
+    tas: xarray.DataArray
+      Mean daily temperature.
+    lat: xarray.DataArray
+      Latitude coordinate.
+    lat_factor: float
+      Latitude factor. Maximum poleward latitude. Default: 75.
+    freq : str
+      Resampling frequency.
+
+    Returns
+    -------
+    xarray.DataArray, [unitless]
+      Latitude Temperature Index.
+
+    Notes
+    -----
+    The latitude factor of `75` is provided for examining the poleward expansion of winegrowing climates under scenarios
+    of climate change. For comparing 20th century/observed historical records, the original scale factor of `60` is more
+    appropriate.
+
+    Let :math:`Tn_{j}` be the average temperature for a given month :math:`j`, :math:`lat_{f}` be the latitude factor,
+    and :math:`lat` be the latitude of the area of interest. Then the Latitude-Temperature Index (:math:`LTI`) is:
+
+    .. math::
+        LTI = max(TN_{j}: j = 1..12)(lat_f - lat)
+
+    References
+    ----------
+    Indice originally published in Jackson, D. I., & Cherry, N. J. (1988). Prediction of a District’s Grape-Ripening
+    Capacity Using a Latitude-Temperature Index (LTI). American Journal of Enology and Viticulture, 39(1), 19‑28.
+
+    Modified latitude factor from Kenny, G. J., & Shao, J. (1992). An assessment of a latitude-temperature index for
+    predicting climate suitability for grapes in Europe. Journal of Horticultural Science, 67(2), 239‑246.
+    https://doi.org/10.1080/00221589.1992.11516243
+    """
+    tas = convert_units_to(tas, "degC")
+
+    tas = tas.resample(time="MS").mean(dim="time")
+    mtwm = tas.resample(time=freq).max(dim="time")
+
+    lat_mask = (abs(lat) >= 0) & (abs(lat) <= lat_factor)
+    lat_coeff = xarray.where(lat_mask, lat_factor - abs(lat), 0)
+
+    lti = mtwm * lat_coeff
+    lti.attrs["units"] = ""
+    return lti
+
+
 @declare_units(
     pr="[precipitation]",
     tasmin="[temperature]",
@@ -341,3 +407,73 @@ def water_budget(
 
     out.attrs["units"] = pr.attrs["units"]
     return out
+
+
+@declare_units(pr="[precipitation]", thresh="[length]")
+def dry_spell_frequency(
+    pr: xarray.DataArray, thresh: str = "1.0 mm", window: int = 3, freq: str = "YS"
+) -> xarray.DataArray:
+    """
+    Return the number of dry periods of n days and more, during which the accumulated precipitation on a window of
+    n days is under the threshold.
+
+    Parameters
+    ----------
+    pr : xarray.DataArray
+      Daily precipitation.
+    thresh : str
+      Accumulated precipitation value under which a period is considered dry.
+    window : int
+      Number of days where the accumulated precipitation is under threshold.
+    freq : str
+      Resampling frequency.
+
+    Returns
+    -------
+    xarray.DataArray
+      The {freq} number of dry periods of minimum {window} days.
+    """
+    pram = rate2amount(pr, out_units="mm")
+    thresh = convert_units_to(thresh, pram)
+
+    out = (
+        (pram.rolling(time=window, center=True).sum() < thresh)
+        .resample(time=freq)
+        .map(rl.windowed_run_events, window=1, dim="time")
+    )
+
+    out.attrs["units"] = ""
+    return out
+
+
+@declare_units(pr="[precipitation]", thresh="[length]")
+def dry_spell_total_length(
+    pr: xarray.DataArray, thresh: str = "1.0 mm", window: int = 3, freq: str = "YS"
+) -> xarray.DataArray:
+    """
+    Return the total number of days in dry periods of n days and more, during which the accumulated precipitation
+    on a window of n days is under the threshold.
+
+    Parameters
+    ----------
+    pr : xarray.DataArray
+      Daily precipitation.
+    thresh : str
+      Accumulated precipitation value under which a period is considered dry.
+    window : int
+      Number of days where the accumulated precipitation is under threshold.
+    freq : str
+      Resampling frequency.
+
+    Returns
+    -------
+    xarray.DataArray
+      The {freq} total number of days in dry periods of minimum {window} days.
+    """
+    pram = rate2amount(pr, out_units="mm")
+    thresh = convert_units_to(thresh, pram)
+
+    mask = pram.rolling(time=window, center=True).sum() < thresh
+    out = (mask.rolling(time=window, center=True).sum() >= 1).resample(time=freq).sum()
+
+    return to_agg_units(out, pram, "count")